+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -571,7 +571,7 @@ class HTTPTool(BaseModel):
     http: HTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6B3TVWQRGGQ37MWVXXVDPCP"
+        "01K6CEZNPR1NY8HMVD9AS1AZR3"
     )
 
     requires_approval: Optional[bool] = False