+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -603,7 +603,7 @@ class RunAgentRequestToolHTTPTool(BaseModel):
     http: RunAgentRequestToolHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6B3TVWYDSSN4R5SCK442FXF"
+        "01K6CEZNPWSZK1M7METM2Y6B8N"
     )
 
     requires_approval: Optional[bool] = False