+++ packages/orq-rc/RELEASES.md
@@ -2999,3 +2999,13 @@ Based on:
 - [python v3.13.0-rc.47] packages/orq-rc
 ### Releases
 - [PyPI v3.13.0-rc.47] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.47 - packages/orq-rc
+
+## 2025-09-30 04:46:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.631.5 (2.716.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.13.0-rc.48] packages/orq-rc
+### Releases
+- [PyPI v3.13.0-rc.48] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.48 - packages/orq-rc